# [aws-ec2-ebs] Do not log "fetched no datapoints" error