--- a/mackerel-plugin-aws-ec2-ebs/lib/aws-ec2-ebs.go
+++ b/mackerel-plugin-aws-ec2-ebs/lib/aws-ec2-ebs.go
@@ -50,55 +50,55 @@ type cloudWatchSetting struct {
 var cloudwatchdefs = map[string](cloudWatchSetting){
 	"ec2.ebs.bandwidth.#.read": cloudWatchSetting{
 		MetricName: "VolumeReadBytes", Statistics: "Sum",
-		CalcFunc:   func(val float64, period float64) float64 { return val / period },
+		CalcFunc: func(val float64, period float64) float64 { return val / period },
 	},
 	"ec2.ebs.bandwidth.#.write": cloudWatchSetting{
 		MetricName: "VolumeWriteBytes", Statistics: "Sum",
-		CalcFunc:   func(val float64, period float64) float64 { return val / period },
+		CalcFunc: func(val float64, period float64) float64 { return val / period },
 	},
 	"ec2.ebs.throughput.#.read": cloudWatchSetting{
 		MetricName: "VolumeReadOps", Statistics: "Sum",
-		CalcFunc:   func(val float64, period float64) float64 { return val / period },
+		CalcFunc: func(val float64, period float64) float64 { return val / period },
 	},
 	"ec2.ebs.throughput.#.write": cloudWatchSetting{
 		MetricName: "VolumeWriteOps", Statistics: "Sum",
-		CalcFunc:   func(val float64, period float64) float64 { return val / period },
+		CalcFunc: func(val float64, period float64) float64 { return val / period },
 	},
 	"ec2.ebs.size_per_op.#.read": cloudWatchSetting{
 		MetricName: "VolumeReadBytes", Statistics: "Average",
-		CalcFunc:   func(val float64, period float64) float64 { return val },
+		CalcFunc: func(val float64, period float64) float64 { return val },
 	},
 	"ec2.ebs.size_per_op.#.write": cloudWatchSetting{
 		MetricName: "VolumeWriteBytes", Statistics: "Average",
-		CalcFunc:   func(val float64, period float64) float64 { return val },
+		CalcFunc: func(val float64, period float64) float64 { return val },
 	},
 	"ec2.ebs.latency.#.read": cloudWatchSetting{
 		MetricName: "VolumeTotalReadTime", Statistics: "Average",
-		CalcFunc:   func(val float64, period float64) float64 { return val * 1000 },
+		CalcFunc: func(val float64, period float64) float64 { return val * 1000 },
 	},
 	"ec2.ebs.latency.#.write": cloudWatchSetting{
 		MetricName: "VolumeTotalWriteTime", Statistics: "Average",
-		CalcFunc:   func(val float64, period float64) float64 { return val * 1000 },
+		CalcFunc: func(val float64, period float64) float64 { return val * 1000 },
 	},
 	"ec2.ebs.queue_length.#.queue_length": cloudWatchSetting{
 		MetricName: "VolumeQueueLength", Statistics: "Average",
-		CalcFunc:   func(val float64, period float64) float64 { return val },
+		CalcFunc: func(val float64, period float64) float64 { return val },
 	},
 	"ec2.ebs.idle_time.#.idle_time": cloudWatchSetting{
 		MetricName: "VolumeIdleTime", Statistics: "Sum",
-		CalcFunc:   func(val float64, period float64) float64 { return val / period * 100 },
+		CalcFunc: func(val float64, period float64) float64 { return val / period * 100 },
 	},
 	"ec2.ebs.throughput_delivered.#.throughput_delivered": cloudWatchSetting{
 		MetricName: "VolumeThroughputPercentage", Statistics: "Average",
-		CalcFunc:   func(val float64, period float64) float64 { return val },
+		CalcFunc: func(val float64, period float64) float64 { return val },
 	},
 	"ec2.ebs.consumed_ops.#.consumed_ops": cloudWatchSetting{
 		MetricName: "VolumeConsumedReadWriteOps", Statistics: "Sum",
-		CalcFunc:   func(val float64, period float64) float64 { return val },
+		CalcFunc: func(val float64, period float64) float64 { return val },
 	},
 	"ec2.ebs.burst_balance.#.burst_balance": cloudWatchSetting{
 		MetricName: "BurstBalance", Statistics: "Average",
-		CalcFunc:   func(val float64, period float64) float64 { return val },
+		CalcFunc: func(val float64, period float64) float64 { return val },
 	},
 }
 
@@ -217,6 +217,8 @@ func (p *EBSPlugin) prepare() error {
 	return nil
 }
 
+var NoDatapointErr = errors.New("fetched no datapoints")
+
 func (p EBSPlugin) getLastPoint(vol *ec2.Volume, metricName string, statType string) (float64, int, error) {
 	now := time.Now()
 
@@ -246,7 +248,7 @@ func (p EBSPlugin) getLastPoint(vol *ec2.Volume, metricName string, statType str
 
 	datapoints := resp.Datapoints
 	if len(datapoints) == 0 {
-		return 0, 0, errors.New("fetched no datapoints")
+		return 0, 0, NoDatapointErr
 	}
 
 	latest := time.Unix(0, 0)
@@ -287,7 +289,7 @@ func (p EBSPlugin) FetchMetrics() (map[string]interface{}, error) {
 				val, period, err := p.getLastPoint(vol, cloudwatchdef.MetricName, cloudwatchdef.Statistics)
 				if err != nil {
 					retErr := errors.New(volumeID + " " + err.Error() + ":" + cloudwatchdef.MetricName)
-					if err.Error() == "fetched no datapoints" {
+					if err == NoDatapointErr {
 						getStderrLogger().Println(retErr)
 					} else {
 						return nil, retErr