--- a/mackerel-plugin-aws-ec2-ebs/lib/aws-ec2-ebs.go
+++ b/mackerel-plugin-aws-ec2-ebs/lib/aws-ec2-ebs.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"flag"
 	"log"
-	"os"
 	"strings"
 	"time"