Some metrics does not have datapoints when the volume is idle, therefore do not output such error to Stderr, to keep clean mackerel-agent.log

BEFORE:
```
$ mackerel-plugin-aws-ec2-ebs
2017/11/16 21:26:13 vol-DUMMY fetched no datapoints:VolumeReadBytes
2017/11/16 21:26:13 vol-DUMMY fetched no datapoints:VolumeReadBytes
2017/11/16 21:26:13 vol-DUMMY fetched no datapoints:VolumeTotalReadTime
ec2.ebs.queue_length.vol-DUMMY.queue_length	0.000100	1510835172
ec2.ebs.idle_time.vol-DUMMY.idle_time	99.990000	1510835172
ec2.ebs.bandwidth.vol-DUMMY.write	3973.120000	1510835172
ec2.ebs.latency.vol-DUMMY.write	0.215827	1510835172
ec2.ebs.burst_balance.vol-DUMMY.burst_balance	100.000000	1510835172
ec2.ebs.throughput.vol-DUMMY.read	0.000000	1510835172
ec2.ebs.throughput.vol-DUMMY.write	0.463333	1510835172
ec2.ebs.size_per_op.vol-DUMMY.write	8575.079137	1510835172
```

AFTER:
```
$ ./mackerel-plugin-aws-ec2-ebs
ec2.ebs.bandwidth.vol-DUMMY.write	3973.120000	1510835159
ec2.ebs.throughput.vol-DUMMY.read	0.000000	1510835159
ec2.ebs.throughput.vol-DUMMY.write	0.463333	1510835159
ec2.ebs.size_per_op.vol-DUMMY.write	8575.079137	1510835159
ec2.ebs.queue_length.vol-DUMMY.queue_length	0.000100	1510835159
ec2.ebs.idle_time.vol-DUMMY.idle_time	99.990000	1510835159
ec2.ebs.burst_balance.vol-DUMMY.burst_balance	100.000000	1510835159
ec2.ebs.latency.vol-DUMMY.write	0.215827	1510835159
```